requirements.txt
fastapi
uvicorn
spotipy
librosa
sentence-transformers
faiss-cpu
streamlit
python-dotenv
requests
gunicorn

